+++ module/Structure/src/Structure/Entity/Db/UniteRecherche.php
@@ -200,4 +200,14 @@ class UniteRecherche implements
     {
         return 'unite_rech';
     }
+
+    /** Pour macro ****************************************************************************************************/
+
+    /**
+     * @noinspection PhpUnusedMethod (il s'agit d'une méthode utilisée par les macros)
+     */
+    public function getSigle(): string
+    {
+        return $this->structure->getSigle() ? $this->structure->getSigle() : "";
+    }
 }